# feat: remove unused optimizations