--- a/.github/workflows/rcBuild.yml
+++ b/.github/workflows/rcBuild.yml
@@ -82,29 +82,3 @@ jobs:
               "npm-package": "sfdx-cli",
               "version": "${{ github.event.release.tag_name }}"
             }
-
-  # failure-notify:
-  #   needs: [pack-verify-upload-tarballs, npm-release, docker-slim, docker-full, pack-upload-win, pack-upload-mac]
-  #   if: always() && failure()
-  #   runs-on: ubuntu-latest
-  #   steps:
-  #     - name: Announce Failure
-  #       id: slack
-  #       uses: slackapi/slack-github-action@v1.21.0
-  #       env:
-  #         SLACK_WEBHOOK_URL: ${{ secrets.CLI_ALERTS_SLACK_WEBHOOK }}
-  #         SLACK_WEBHOOK_TYPE: INCOMING_WEBHOOK
-  #       with:
-  #         payload: |
-  #           {
-  #             "text": "RC fail",
-  #             "blocks": [
-  #               {
-  #                 "type": "section",
-  #                 "text": {
-  #                   "type": "mrkdwn",
-  #                   "text": "GitHub Action build result: ${{ job.status }}\n${{ github.event.pull_request.html_url || github.event.head_commit.url }}"
-  #                 }
-  #               }
-  #             ]
-  #           }