… package.json
… src/hooks/lazyRequire.ts
… src/lazyRequire.ts
… test/lazyRequire.test.ts
… yarn.lock