### What does this PR do?
remove 2 dependencies (lazyRequire, requireAnalytics)

why?
- they were experimental
- they required users to opt in via an env we've not publicized
- `sf` is not using these
- that's 2 repos we don't have to  maintain/migrate 

### What issues does this PR fix or reference?
@W-6012451@