@shetzel 2022-09-21
Testing with and without these changes show very little difference when lazy loading is in use, at least on my machine.

**With lazy loading:**
```
hyperfine --warmup 5 './bin/run version'
Benchmark 1: ./bin/run version
  Time (mean ± σ):     479.7 ms ±   3.1 ms    [User: 497.8 ms, System: 85.6 ms]
  Range (min … max):   474.5 ms … 483.2 ms    10 runs

hyperfine --warmup 5 './bin/run commands --help'
Benchmark 1: ./bin/run commands --help
  Time (mean ± σ):     450.4 ms ±   2.8 ms    [User: 473.4 ms, System: 76.1 ms]
  Range (min … max):   447.2 ms … 455.8 ms    10 runs
```

**Without lazy loading:**
```
hyperfine --warmup 5 './bin/run version'
Benchmark 1: ./bin/run version
  Time (mean ± σ):     500.2 ms ±   7.2 ms    [User: 510.9 ms, System: 92.4 ms]
  Range (min … max):   493.4 ms … 513.7 ms    10 runs

hyperfine --warmup 5 './bin/run commands --help'
Benchmark 1: ./bin/run commands --help
  Time (mean ± σ):     455.4 ms ±   3.0 ms    [User: 476.1 ms, System: 77.1 ms]
  Range (min … max):   451.6 ms … 462.6 ms    10 runs
```